Update kaprekarUtils.c
Changed return false to return 0 because that is the way to return false about a statement. Changed log10(n) to log10(square) because it is computing the digits of the square, not the digits itself. fixed spelling errors of variables so they are declared (numberOfDigits to numDigits, modulous to modulus). Changed long modulus = 0 to modulus =1 because modulus is not storing a number. changed n = first + second to n == first + second because it has to equaling to first + second not assigning its value.